requirements.txt
langgraph==0.0.19
uuid==1.30
langchain>=0.1.0
langchain-core>=0.1.0
langchain-aws>=0.1.0
boto3>=1.34.0
python-dotenv>=1.0.0 